DCS-839 moving code managing state into shared location so it can be reused
